Async networking: Split declaration and definition inside header
